2010-10-31  Tatsuhiro Tsujikawa  <t-tujikawa@users.sourceforge.net>
	Fixed the bug that downloading > 4GB file fails on 32bit systems.
	* src/HttpRequest.cc
	* src/HttpRequestCommand.cc

+++ src/HttpRequestCommand.cc
@@ -188,7 +188,8 @@ bool HttpRequestCommand::executeInternal() {
               getPieceStorage()->getNextUsedIndex(segment->getIndex());
             endOffset = std::min
               (static_cast<off_t>(getFileEntry()->getLength()),
-               getFileEntry()->gtoloff(segment->getSegmentLength()*nextIndex));
+               getFileEntry()->gtoloff
+               (static_cast<off_t>(segment->getSegmentLength())*nextIndex));
           }
           SharedHandle<HttpRequest> httpRequest
             (createHttpRequest(getRequest(),
